DEV: Migrate jump-to-post to the new modal api
