+++ app/assets/javascripts/discourse/app/components/modal/jump-to-post.hbs
@@ -1,0 +1,51 @@
+<DModal
+  @closeModal={{@closeModal}}
+  @title={{i18n "topic.progress.jump_prompt_long"}}
+  class="jump-to-post-modal"
+>
+  <:body>
+    <div class="jump-to-post-form">
+      <div class="jump-to-post-control">
+        <span class="index">#</span>
+        <Input
+          @value={{this.postNumber}}
+          @type="number"
+          autofocus="true"
+          id="post-jump"
+        />
+        <span class="input-hint-text post-number">
+          {{i18n "topic.progress.jump_prompt_of" count=this.filteredPostsCount}}
+        </span>
+      </div>
+
+      <div class="separator">
+        <span class="text">
+          {{i18n "topic.progress.jump_prompt_or"}}
+        </span>
+        <hr class="right" />
+      </div>
+
+      <div class="jump-to-date-control">
+        <span class="input-hint-text post-date">
+          {{i18n "topic.progress.jump_prompt_to_date"}}
+        </span>
+        <DatePicker
+          @value={{this.postDate}}
+          @defaultDate="YYYY-MM-DD"
+          id="post-date"
+          class="date-input"
+        />
+      </div>
+    </div>
+  </:body>
+
+  <:footer>
+    <DButton
+      @action={{this.jump}}
+      @label="composer.modal_ok"
+      type="submit"
+      class="btn-primary"
+    />
+    <DModalCancel @close={{@closeModal}} />
+  </:footer>
+</DModal>
