--- app/assets/javascripts/discourse/app/templates/modal/jump-to-post.hbs
@@ -1,46 +1,0 @@
-<DModalBody @title="topic.progress.jump_prompt_long">
-
-  <div class="jump-to-post-form">
-    <div class="jump-to-post-control">
-      <span class="index">#</span>
-      <Input
-        id="post-jump"
-        @type="number"
-        @value={{this.postNumber}}
-        @insert-newline={{action "jump"}}
-        autofocus="true"
-      />
-      <span class="input-hint-text post-number">
-        {{i18n "topic.progress.jump_prompt_of" count=this.filteredPostsCount}}
-      </span>
-    </div>
-
-    <div class="separator">
-      <span class="text">
-        {{i18n "topic.progress.jump_prompt_or"}}
-      </span>
-      <hr class="right" />
-    </div>
-
-    <div class="jump-to-date-control">
-      <span class="input-hint-text post-date">
-        {{i18n "topic.progress.jump_prompt_to_date"}}
-      </span>
-      <DatePicker
-        @id="post-date"
-        @class="date-input"
-        @value={{this.postDate}}
-        @defaultDate="YYYY-MM-DD"
-      />
-    </div>
-  </div>
-</DModalBody>
-
-<div class="modal-footer">
-  <DButton
-    @class="btn-primary"
-    @action={{action "jump"}}
-    @label="composer.modal_ok"
-  />
-  <DModalCancel @close={{route-action "closeModal"}} />
-</div>
