+++ app/assets/javascripts/discourse/tests/acceptance/jump-to-test.js
@@ -1,4 +1,4 @@
-import { acceptance, exists } from "discourse/tests/helpers/qunit-helpers";
+import { acceptance } from "discourse/tests/helpers/qunit-helpers";
 import { click, currentURL, fillIn, visit } from "@ember/test-helpers";
 import { test } from "qunit";
 
@@ -7,11 +7,11 @@ acceptance("Jump to", function (needs) {
   needs.mobileView();
 
   needs.pretender((server, helper) => {
-    server.get("/t/280/excerpts.json", () => helper.response(200, []));
-    server.get("/t/280/3.json", () => helper.response(200, {}));
+    server.get("/t/280/excerpts.json", () => helper.response([]));
+    server.get("/t/280/3.json", () => helper.response({}));
     server.get("/posts/by-date/280/:date", (req) => {
       if (req.params["date"] === "2014-02-24") {
-        return helper.response(200, {
+        return helper.response({
           post_number: 3,
         });
       }
@@ -25,7 +25,7 @@ acceptance("Jump to", function (needs) {
     await click("nav#topic-progress .nums");
     await click("button.jump-to-post");
 
-    assert.ok(exists(".jump-to-post-modal"), "it shows the modal");
+    assert.dom(".jump-to-post-modal").exists("it shows the modal");
 
     await fillIn("input.date-picker", "2014-02-24");
     await click(".jump-to-post-modal .btn-primary");
